% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape.R
\name{get_play_types}
\alias{get_play_types}
\title{Get Play Type Stats}
\usage{
get_play_types(pull_type = "team", season_yr = 2018, limit = 500,
  names = "offensive", seasonType = "reg")
}
\value{
tibble with data compiled for all play type categories
}
\description{
Wraps around nbaTools::GetPlayTypeStats to scrape data off stas.nba.com
  and then cleans up the data types, values, and names to be consistent
  with goals of bballfanalytics
}
\examples{
tbl <- get_play_types(pull_type = "player", seasonType = "post")
tbl <- get_play_types(pull_type = "team", names = "defensive")
}
\keyword{play}
\keyword{player}
\keyword{team}
\keyword{type}
